rename gist filename: 'nopaste.txt' to 'nopaste'
The default filename for the gist API, 'nopaste.txt' (e.g. when reading
from STDIN) always comes out without syntax highlighting, due to the new
gist using the filename extension.

If we rename the filename to 'nopaste', without an extension, it uses
the shebang line, if available, to choose a syntax highlighting scheme.